+++ mars/services/web/worker.py
@@ -12,5 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+
 async def start(*_, **__):  # pragma: no cover
     return
